+++ .solcover.ts
@@ -1,4 +1,0 @@
-module.exports = {
-  skipFiles: ["interfaces/", "mock/", "libs/crypto/ECDSA512.sol", "libs/crypto/ECDSA384.sol"],
-  configureYulOptimizer: true,
-};