+++ test/contracts-registry/presets/MultiOwnableContractsRegistry.test.ts
@@ -1,31 +1,25 @@
-import { ethers } from "hardhat";
 import { expect } from "chai";
+import hre from "hardhat";
 
-import { SignerWithAddress } from "@nomicfoundation/hardhat-ethers/signers";
-
-import { Reverter } from "@/test/helpers/reverter";
+import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/types";
 
 import { MultiOwnableContractsRegistry } from "@ethers-v6";
 
-describe("MultiOwnableContractsRegistry", () => {
-  const reverter = new Reverter();
+const { ethers } = await hre.network.connect();
 
-  let SECOND: SignerWithAddress;
+describe("MultiOwnableContractsRegistry", () => {
+  let SECOND: HardhatEthersSigner;
 
   let contractsRegistry: MultiOwnableContractsRegistry;
 
-  before("setup", async () => {
+  beforeEach("setup", async () => {
     [, SECOND] = await ethers.getSigners();
 
     const MultiOwnableContractsRegistry = await ethers.getContractFactory("MultiOwnableContractsRegistry");
     contractsRegistry = await MultiOwnableContractsRegistry.deploy();
 
     await contractsRegistry.__MultiOwnableContractsRegistry_init();
-
-    await reverter.snapshot();
   });
-
-  afterEach(reverter.revert);
 
   describe("access", () => {
     it("should not initialize twice", async () => {
@@ -75,23 +69,23 @@ describe("MultiOwnableContractsRegistry", () => {
 
   describe("coverage", () => {
     it("should call these methods", async () => {
-      await expect(contractsRegistry.injectDependencies("")).to.be.reverted;
+      await expect(contractsRegistry.injectDependencies("")).to.be.revert(ethers);
 
-      await expect(contractsRegistry.injectDependenciesWithData("", "0x")).to.be.reverted;
+      await expect(contractsRegistry.injectDependenciesWithData("", "0x")).to.be.revert(ethers);
 
-      await expect(contractsRegistry.upgradeContract("", ethers.ZeroAddress)).to.be.reverted;
+      await expect(contractsRegistry.upgradeContract("", ethers.ZeroAddress)).to.be.revert(ethers);
 
-      await expect(contractsRegistry.upgradeContractAndCall("", ethers.ZeroAddress, "0x")).to.be.reverted;
+      await expect(contractsRegistry.upgradeContractAndCall("", ethers.ZeroAddress, "0x")).to.be.revert(ethers);
 
-      await expect(contractsRegistry.addContract("", ethers.ZeroAddress)).to.be.reverted;
+      await expect(contractsRegistry.addContract("", ethers.ZeroAddress)).to.be.revert(ethers);
 
-      await expect(contractsRegistry.addProxyContract("", ethers.ZeroAddress)).to.be.reverted;
+      await expect(contractsRegistry.addProxyContract("", ethers.ZeroAddress)).to.be.revert(ethers);
 
-      await expect(contractsRegistry.addProxyContractAndCall("", ethers.ZeroAddress, "0x")).to.be.reverted;
+      await expect(contractsRegistry.addProxyContractAndCall("", ethers.ZeroAddress, "0x")).to.be.revert(ethers);
 
-      await expect(contractsRegistry.justAddProxyContract("", ethers.ZeroAddress)).to.be.reverted;
+      await expect(contractsRegistry.justAddProxyContract("", ethers.ZeroAddress)).to.be.revert(ethers);
 
-      await expect(contractsRegistry.removeContract("")).to.be.reverted;
+      await expect(contractsRegistry.removeContract("")).to.be.revert(ethers);
     });
   });
 });