+++ test/diamond/DiamondERC165.test.ts
@@ -1,20 +1,15 @@
-import { ethers } from "hardhat";
 import { expect } from "chai";
+import hre from "hardhat";
 
-import { Reverter } from "@/test/helpers/reverter";
-import { getSelectors, FacetAction } from "@/test/helpers/diamond-helper";
+import { FacetAction, getSelectors } from "@test-helpers";
 
-import { OwnableDiamondMock, DiamondERC165, Diamond } from "@ethers-v6";
+import { Diamond, DiamondERC165, OwnableDiamondMock } from "@ethers-v6";
+
+const { ethers } = await hre.network.connect();
 
 describe("DiamondERC165", () => {
-  const reverter = new Reverter();
-
   let erc165: DiamondERC165;
   let diamond: OwnableDiamondMock;
-
-  before("setup", async () => {
-    await reverter.snapshot();
-  });
 
   beforeEach("setup", async () => {
     const OwnableDiamond = await ethers.getContractFactory("OwnableDiamondMock");
@@ -34,10 +29,8 @@ describe("DiamondERC165", () => {
     await diamond.__OwnableDiamondMock_init();
     await diamond.diamondCutShort(facets);
 
-    erc165 = <DiamondERC165>DiamondERC165.attach(await diamond.getAddress());
+    erc165 = DiamondERC165.attach(await diamond.getAddress());
   });
-
-  afterEach(reverter.revert);
 
   describe("ERC165", () => {
     it("should support IERC165", async () => {