+++ test/libs/utils/EndianConverter.test.ts
@@ -1,12 +1,13 @@
-import { EndianConverterMock } from "@/generated-types/ethers";
-import { reverseBytes } from "@/test/helpers/bytes-helpers";
-import { Reverter } from "@/test/helpers/reverter";
 import { expect } from "chai";
-import { ethers } from "hardhat";
+import hre from "hardhat";
+
+import { reverseBytes } from "@test-helpers";
+
+import { EndianConverterMock } from "@ethers-v6";
+
+const { ethers } = await hre.network.connect();
 
 describe("EndianConverter", () => {
-  const reverter = new Reverter();
-
   let converter: EndianConverterMock;
 
   const bytes2 = "0x13d4";
@@ -15,14 +16,10 @@ describe("EndianConverter", () => {
   const bytes16 = "0x00100513fde409ababcdef1234567890";
   const bytes32 = "0x123456789abcdef000112233445566778899aabbccddeefffa000513fdd409ab";
 
-  before(async () => {
+  beforeEach("setup", async () => {
     const EndianConverterMock = await ethers.getContractFactory("EndianConverterMock");
     converter = await EndianConverterMock.deploy();
-
-    await reverter.snapshot();
   });
-
-  afterEach(reverter.revert);
 
   it("should correctly convert bytes", async () => {
     let expected = reverseBytes(bytes2);