+++ test/utils/BlockGuard.test.ts
@@ -1,31 +1,29 @@
-import { ethers } from "hardhat";
 import { expect } from "chai";
+import hre from "hardhat";
 
-import { SignerWithAddress } from "@nomicfoundation/hardhat-ethers/signers";
-import { time } from "@nomicfoundation/hardhat-network-helpers";
-
-import { Reverter } from "@/test/helpers/reverter";
+import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/types";
+import { Time } from "@nomicfoundation/hardhat-network-helpers/types";
 
 import { BlockGuardMock } from "@ethers-v6";
 
-describe("BlockGuard", () => {
-  const reverter = new Reverter();
+const { ethers, networkHelpers } = await hre.network.connect();
 
-  let FIRST: SignerWithAddress;
-  let SECOND: SignerWithAddress;
+describe("BlockGuard", () => {
+  let time: Time;
+
+  let FIRST: HardhatEthersSigner;
+  let SECOND: HardhatEthersSigner;
 
   let mock: BlockGuardMock;
 
-  before("setup", async () => {
+  beforeEach("setup", async () => {
+    time = networkHelpers.time;
+
     [FIRST, SECOND] = await ethers.getSigners();
 
     const BlockGuardMock = await ethers.getContractFactory("BlockGuardMock");
     mock = await BlockGuardMock.deploy();
-
-    await reverter.snapshot();
   });
-
-  afterEach(reverter.revert);
 
   describe("lockBlock", () => {
     it("should return zero if the resource key hasn't been locked", async () => {